Fix CSI health server
